move common methods to helper.hpp
